fix(console): application guide drawer mode in details page
