+++ packages/console/src/pages/Applications/components/Guide/index.tsx
@@ -5,13 +5,11 @@ import i18next from 'i18next';
 import type { MDXProps } from 'mdx/types';
 import type { LazyExoticComponent } from 'react';
 import { useEffect, useContext, cloneElement, lazy, Suspense, useState } from 'react';
-import Modal from 'react-modal';
 
 import CodeEditor from '@/components/CodeEditor';
 import TextLink from '@/components/TextLink';
 import { AppEndpointsContext } from '@/contexts/AppEndpointsProvider';
 import DetailsSummary from '@/mdx-components/DetailsSummary';
-import * as modalStyles from '@/scss/modal.module.scss';
 import type { SupportedSdk } from '@/types/applications';
 import { applicationTypeAndSdkTypeMappings } from '@/types/applications';
 
@@ -23,7 +21,7 @@ import * as styles from './index.module.scss';
 type Props = {
   app?: Application;
   isCompact?: boolean;
-  onClose: (id: string) => void;
+  onClose: () => void;
 };
 
 const Guides: Record<string, LazyExoticComponent<(props: MDXProps) => JSX.Element>> = {
@@ -72,73 +70,62 @@ const Guide = ({ app, isCompact, onClose }: Props) => {
   const guideI18nKey = `${selectedSdk}_${locale}`.toLowerCase();
   const GuideComponent = Guides[guideI18nKey] ?? Guides[selectedSdk.toLowerCase()];
 
-  const closeModal = () => {
-    onClose(appId);
-  };
-
   return (
-    <Modal
-      shouldCloseOnEsc
-      isOpen={Boolean(app)}
-      className={modalStyles.fullScreen}
-      onRequestClose={closeModal}
-    >
-      <div className={styles.container}>
-        <GuideHeader
-          appName={appName}
-          selectedSdk={selectedSdk}
-          isCompact={isCompact}
-          onClose={closeModal}
-        />
-        <div className={styles.content}>
-          {cloneElement(<SdkSelector sdks={sdks} selectedSdk={selectedSdk} />, {
-            className: styles.banner,
-            isCompact,
-            onChange: setSelectedSdk,
-            onToggle: () => {
-              setActiveStepIndex(0);
-            },
-          })}
-          <MDXProvider
-            components={{
-              code: ({ className, children }) => {
-                const [, language] = /language-(\w+)/.exec(className ?? '') ?? [];
+    <div className={styles.container}>
+      <GuideHeader
+        appName={appName}
+        selectedSdk={selectedSdk}
+        isCompact={isCompact}
+        onClose={onClose}
+      />
+      <div className={styles.content}>
+        {cloneElement(<SdkSelector sdks={sdks} selectedSdk={selectedSdk} />, {
+          className: styles.banner,
+          isCompact,
+          onChange: setSelectedSdk,
+          onToggle: () => {
+            setActiveStepIndex(0);
+          },
+        })}
+        <MDXProvider
+          components={{
+            code: ({ className, children }) => {
+              const [, language] = /language-(\w+)/.exec(className ?? '') ?? [];
 
-                return language ? (
-                  <CodeEditor isReadonly language={language} value={String(children)} />
-                ) : (
-                  <code>{String(children)}</code>
-                );
-              },
-              a: ({ children, ...props }) => (
-                <TextLink {...props} target="_blank" rel="noopener noreferrer">
-                  {children}
-                </TextLink>
-              ),
-              details: DetailsSummary,
-            }}
-          >
-            <Suspense fallback={<StepsSkeleton />}>
-              {GuideComponent && (
-                <GuideComponent
-                  appId={appId}
-                  appSecret={appSecret}
-                  endpoint={userEndpoint}
-                  redirectUris={oidcClientMetadata.redirectUris}
-                  postLogoutRedirectUris={oidcClientMetadata.postLogoutRedirectUris}
-                  activeStepIndex={activeStepIndex}
-                  isCompact={isCompact}
-                  onNext={(nextIndex: number) => {
-                    setActiveStepIndex(nextIndex);
-                  }}
-                  onComplete={closeModal}
-                />
-              )}
-            </Suspense>
-          </MDXProvider>
-        </div>
+              return language ? (
+                <CodeEditor isReadonly language={language} value={String(children)} />
+              ) : (
+                <code>{String(children)}</code>
+              );
+            },
+            a: ({ children, ...props }) => (
+              <TextLink {...props} target="_blank" rel="noopener noreferrer">
+                {children}
+              </TextLink>
+            ),
+            details: DetailsSummary,
+          }}
+        >
+          <Suspense fallback={<StepsSkeleton />}>
+            {GuideComponent && (
+              <GuideComponent
+                appId={appId}
+                appSecret={appSecret}
+                endpoint={userEndpoint}
+                redirectUris={oidcClientMetadata.redirectUris}
+                postLogoutRedirectUris={oidcClientMetadata.postLogoutRedirectUris}
+                activeStepIndex={activeStepIndex}
+                isCompact={isCompact}
+                onNext={(nextIndex: number) => {
+                  setActiveStepIndex(nextIndex);
+                }}
+                onComplete={onClose}
+              />
+            )}
+          </Suspense>
+        </MDXProvider>
       </div>
-    </Modal>
+    </div>
   );
 };
